requirements.txt
flask
openai
python-dotenv
python-docx